Organization pagination addition and user pagination location update
