+++ schema/organization/organization.graphql
@@ -57,5 +57,50 @@ module.exports = `
     organization: Organization!
   }
 
+  input OrganizationWhereInput {
+    id: ID
+    id_not: ID
+    id_in: [ID!]
+    id_not_in: [ID!]
+    id_contains: ID
+    id_starts_with: ID
+
+    name: String
+    name_not: String
+    name_in: [String!]
+    name_not_in: [String!]
+    name_contains: String
+    name_starts_with: String
+
+    description: String
+    description_not: String
+    description_in: [String!]
+    description_not_in: [String!]
+    description_contains: String
+    description_starts_with: String
+
+    apiUrl: String
+    apiUrl_not: String
+    apiUrl_in: [String!]
+    apiUrl_not_in: [String!]
+    apiUrl_contains: String
+    apiUrl_starts_with: String
+
+    visibleInSearch: Boolean
+
+    isPublic: Boolean
+   }
+
+   enum OrganizationOrderByInput {
+    id_ASC
+    id_DESC
+    name_ASC
+    name_DESC
+    description_ASC
+    description_DESC
+    apiUrl_ASC
+    apiUrl_DESC
+   }
+
 
 `
